RF_05.002.01 | Multibranch Pipeline > combining all Multibranch Pipeline -related classes to one class (MultibranchPipelineTest)
